requirements.txt
accelerate==1.0.1
backports.tarfile==1.2.0
bigtree==1.0.0
dash==3.3.0
decord==0.6.0
diffusers==0.31.0
einops==0.8.1
en-core-web-sm==3.8.0
ffmpeg-python==0.2.0
google-genai==1.3.0
google-generativeai==0.8.3
grpcio-status==1.69.0
h5py==3.15.1
imps==0.2.6
jaraco.collections==5.1.0
kagglehub==0.3.13
levenshtein==0.27.1
lora==0.3.0
moviepy==2.1.2
mplcursors==0.7
musical-mel-transform==0.1.1
nltk==3.9.1
nnaudio==0.3.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pip-chill==1.0.3
pyaudio==0.2.14
pychord==1.2.2
pydub==0.25.1
pyqt6-webengine==6.10.0
roman-numerals==3.1.0
serial==0.0.97
sounddevice==0.5.3
spacy==3.8.11
speechrecognition==3.11.0
tomli==2.0.1
torchvision==0.22.0+cu128
transformers==4.47.1
vsrife==5.2.0
wordcloud==1.9.4
